split out cookie notice text classifiers
