+++ src/lib.mjs
@@ -23,6 +23,8 @@ import { KnownDevices } from 'puppeteer-core'
 import { puppeteerConfigForArgs } from './puppeteer.mjs'
 import { templateProfilePathForArgs, parseListCatalogComponentIds, isValidChromeComponentId, isKeeplistedComponentId, getExtensionVersion, getOptionalDefaultComponentIds, replaceVersion, toggleAdblocklists, proxyUrlWithAuth, checkAllComponentsRegistered, fixupBundleStackTrace, getBundlePaths } from './util.mjs'
 
+import { cookieNoticeClassifier } from './text-classification.mjs'
+
 // Generate a random string between [a000000000, zzzzzzzzzz] (base 36)
 const generateRandomToken = () => {
   const min = Number.parseInt('a000000000', 36)
@@ -35,85 +37,12 @@ const openai = new OpenAI({
   apiKey: process.env.OPENAI_API_KEY || 'ollama'
 })
 
-const MAX_LENGTH = 500
-
-const keywordClassifierFallback = (innerText) => {
-  const keywords = [
-    'cookies',
-    'consent',
-    'privacy',
-    'analytics',
-    'accept',
-    'only necessary',
-    'reject'
-  ]
-  const lower = innerText.toLowerCase()
-  for (const keyword of keywords) {
-    if (lower.includes(keyword)) {
-      return true
-    }
-  }
-  return false
-}
-
 const inPageAPI = {
   extractFrameText: (iframeHandle) => {
     return iframeHandle.contentFrame().then(frameDoc => frameDoc.evaluate(() => document.documentElement.innerText))
   },
-  classifyInnerText: (innerText) => {
-    let innerTextSnippet = innerText.slice(0, MAX_LENGTH)
-    let ifTruncated = ''
-    if (innerTextSnippet.length !== innerText.length) {
-      innerTextSnippet += '...'
-      ifTruncated = `the first ${MAX_LENGTH} characters of `
-    }
-    const systemPrompt = `Your task is to classify text from the innerText property of HTML overlay elements.
-
-An overlay element is considered to be a "cookie consent notice" if it meets all of these criteria:
-1. it explicitly notifies the user of the site's use of cookies or other storage technology, such as: "We use cookies...", "This site uses...", etc.
-2. it offers the user choices for the usage of cookies on the site, such as: "Accept", "Reject", "Learn More", etc., or informs the user that their use of the site means they accept the usage of cookies.
-
-Note: This definition does not include adult content notices or any other type of notice that is primarily focused on age verification or content restrictions. Cookie consent notices are specifically intended to inform users about the website's use of cookies and obtain their consent for such use.
-
-Note: A cookie consent notice should specifically relate to the site's use of cookies or other storage technology that stores data on the user's device, such as HTTP cookies, local storage, or session storage. Requests for permission to access geolocation information, camera, microphone, etc., do not fall under this category.
-
-Note: Do NOT classify a website header or footer as a "cookie consent notice". Website headers or footers may contain a list of links, possibly including a privacy policy, cookie policy, or terms of service document, but their primary purpose is navigational rather than informational.
-`
-    const prompt = `
-The following text was captured from ${ifTruncated}the innerText of an HTML overlay element:
-
-\`\`\`
-${innerTextSnippet}
-\`\`\`
-
-Is the overlay element above considered to be a "cookie consent notice"? Provide your answer as a boolean.
-`
-    return openai.chat.completions.create({
-      model: process.env.OPENAI_MODEL || 'llama3',
-      messages: [
-        { role: 'system', content: systemPrompt },
-        { role: 'user', content: prompt }
-      ],
-      // We only need enough tokens for "true" or "false"
-      max_tokens: 2,
-      // Fixed seed and zero temperature to avoid randomized responses
-      seed: 1,
-      temperature: 0,
-      // Only consider the single most likely next token
-      top_p: 0
-    }).then(response => {
-      const answer = response.choices[0].message.content
-      return {
-        classifier: 'llm',
-        classification: answer.match(/true/i)
-      }
-    }).catch(e => {
-      console.error('LLM classification failed:', e)
-      return {
-        classifier: 'keyword',
-        classification: keywordClassifierFallback(innerText)
-      }
-    })
+  classifyCookieNoticeText: (innerText) => {
+    return cookieNoticeClassifier(innerText, openai)
   },
   getETLDP1: (() => {
     let init
